fix get_time_ms on windows

+++ src/platforms/pc/platform.c
@@ -86,7 +86,7 @@ uint64_t get_time_ms() {
   struct timeval tv;
   gettimeofday(&tv, NULL);
 
-  return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+  return (uint64_t)(tv.tv_sec) * 1000 + (uint64_t)(tv.tv_usec) / 1000;
 }
 
 void sleep_ms(uint32_t ms) {
